Fixes and improvements for PnT code

+++ src/ditl/sim/pnt/PrioNewest.java
@@ -8,7 +8,7 @@ public class PrioNewest extends ArrivalWho {
 
 	@Override
 	public Integer whoToPush(Message msg, Set<Integer> infected, Set<Integer> sane){
-		double newestArrival = Double.MIN_VALUE;
+		long newestArrival = Long.MIN_VALUE;
 		for( Integer i : sane ){
 			if ( _arrivalTimes.get(i) > newestArrival ){
 				newestArrival = _arrivalTimes.get(i);
